+++ .github/pull_request_template.md
@@ -18,15 +18,15 @@
 - 뷰데이터와 엔티티데이터의 분리에 대한 이해
 - entities -> features -> UI 계층에 대한 이해
 
-- [ ] Component에서 사용되는 Data가 아닌 로직들은 hook으로 옮겨졌나요?
-- [ ] 주어진 hook의 책임에 맞도록 코드가 분리가 되었나요?
-- [ ] 계산함수는 순수함수로 작성이 되었나요?
-- [ ] Component에서 사용되는 Data가 아닌 로직들은 hook으로 옮겨졌나요?
-- [ ] 주어진 hook의 책임에 맞도록 코드가 분리가 되었나요?
-- [ ] 계산함수는 순수함수로 작성이 되었나요?
-- [ ] 특정 Entitiy만 다루는 함수는 분리되어 있나요?
-- [ ] 특정 Entitiy만 다루는 Component와 UI를 다루는 Component는 분리되어 있나요?
-- [ ] 데이터 흐름에 맞는 계층구조를 이루고 의존성이 맞게 작성이 되었나요?
+- [x] Component에서 사용되는 Data가 아닌 로직들은 hook으로 옮겨졌나요?
+- [x] 주어진 hook의 책임에 맞도록 코드가 분리가 되었나요?
+- [x] 계산함수는 순수함수로 작성이 되었나요?
+- [x] Component에서 사용되는 Data가 아닌 로직들은 hook으로 옮겨졌나요?
+- [x] 주어진 hook의 책임에 맞도록 코드가 분리가 되었나요?
+- [x] 계산함수는 순수함수로 작성이 되었나요?
+- [x] 특정 Entitiy만 다루는 함수는 분리되어 있나요?
+- [x] 특정 Entitiy만 다루는 Component와 UI를 다루는 Component는 분리되어 있나요?
+- [x] 데이터 흐름에 맞는 계층구조를 이루고 의존성이 맞게 작성이 되었나요?
 
 ### 심화과제
 
@@ -35,7 +35,7 @@
 - 재사용 가능한 Custom 유틸 함수를 만들어 보기
 - 그래서 엔티티와는 어떤 다른 계층적 특징을 가지는지 이해하기
 
-- [ ] UI 컴포넌트 계층과 엔티티 컴포넌트의 계층의 성격이 다르다는 것을 이해하고 적용했는가?
+- [x] UI 컴포넌트 계층과 엔티티 컴포넌트의 계층의 성격이 다르다는 것을 이해하고 적용했는가?
 - [ ] 엔티티 Hook과 라이브러리 훅과의 계층의 성격이 다르다는 것을 이해하고 적용했는가?
 - [ ] 엔티티 순수함수와 유틸리티 함수의 계층의 성격이 다르다는 것을 이해하고 적용했는가?
